Fix system_updates_checks interval fields type usage

+++ pkg/monitoring/updates/updates.go
@@ -29,14 +29,14 @@ func Shutdown() {
 	}
 }
 
-func GetWatcher(fetchTimeout, checkInterval uint) *Watcher {
+func GetWatcher(fetchTimeout, checkInterval uint32) *Watcher {
 	if watcher != nil {
 		return watcher
 	}
 
 	watcher = &Watcher{
-		fetchTimeout:    time.Duration(uint(time.Second) * fetchTimeout),
-		checkInterval:   time.Duration(uint(time.Second) * checkInterval),
+		fetchTimeout:    time.Duration(int64(time.Second) * int64(fetchTimeout)),
+		checkInterval:   time.Duration(int64(time.Second) * int64(checkInterval)),
 		lastFetchedInfo: nil,
 		interruptChan:   make(chan struct{}),
 	}
